+++ controller/main_donor_list.php
@@ -96,9 +96,6 @@ class main_donor_list extends main_controller
 
 		$this->pagination->generate_template_pagination($pagination_url, 'pagination', 'start', $total_donors, (int) $this->config['topics_per_page'], $start);
 
-		// Get default currency data from the database
-		$default_currency_data = $this->ppde_actions_currency->get_default_currency_data((int) $this->config['ppde_default_currency']);
-
 		// Assign vars to the template
 		$this->template->assign_vars([
 			'L_PPDE_DONORLIST_TITLE' => $this->language->lang('PPDE_DONORLIST_TITLE'),
@@ -110,9 +107,10 @@ class main_donor_list extends main_controller
 
 		// Adds fields to the table schema needed by entity->import()
 		$donorlist_table_schema = [
-			'item_amount'     => ['name' => 'amount', 'type' => 'float'],
-			'item_max_txn_id' => ['name' => 'max_txn_id', 'type' => 'integer'],
-			'item_user_id'    => ['name' => 'user_id', 'type' => 'integer'],
+			'item_amount'      => ['name' => 'amount', 'type' => 'float'],
+			'item_max_txn_id'  => ['name' => 'max_txn_id', 'type' => 'integer'],
+			'item_user_id'     => ['name' => 'user_id', 'type' => 'integer'],
+			'item_mc_currency' => ['name' => 'mc_currency', 'type' => 'string'],
 		];
 
 		$sql_donorlist_ary = $this->ppde_operator_transactions->sql_donorlist_ary(true, $order_by);
@@ -128,11 +126,12 @@ class main_donor_list extends main_controller
 		{
 			$get_last_transaction_sql_ary = $this->ppde_operator_transactions->sql_last_donation_ary($data['max_txn_id']);
 			$last_donation_data = $this->ppde_entity_transactions->get_data($this->ppde_operator_transactions->build_sql_donorlist_data($get_last_transaction_sql_ary), $last_donation_table_schema, 0, 0, true);
+			$currency_mc_data = $this->ppde_actions_currency->get_currency_data($last_donation_data[0]['mc_currency']);
 			$this->template->assign_block_vars('donorrow', [
 				'PPDE_DONOR_USERNAME'       => $this->user_loader->get_username($data['user_id'], 'full', false, false, true),
-				'PPDE_LAST_DONATED_AMOUNT'  => $this->ppde_actions_currency->currency_on_left($last_donation_data[0]['mc_gross'], $default_currency_data[0]['currency_symbol'], (bool) $default_currency_data[0]['currency_on_left']),
+				'PPDE_LAST_DONATED_AMOUNT'  => $this->ppde_actions_currency->format_currency($last_donation_data[0]['mc_gross'], $currency_mc_data[0]['currency_iso_code'], $currency_mc_data[0]['currency_symbol'], (bool) $currency_mc_data[0]['currency_on_left']),
 				'PPDE_LAST_PAYMENT_DATE'    => $this->user->format_date($last_donation_data[0]['payment_date']),
-				'PPDE_TOTAL_DONATED_AMOUNT' => $this->ppde_actions_currency->currency_on_left($data['amount'], $default_currency_data[0]['currency_symbol'], (bool) $default_currency_data[0]['currency_on_left']),
+				'PPDE_TOTAL_DONATED_AMOUNT' => $this->ppde_actions_currency->format_currency($data['amount'], $currency_mc_data[0]['currency_iso_code'], $currency_mc_data[0]['currency_symbol'], (bool) $currency_mc_data[0]['currency_on_left']),
 			]);
 		}