+++ migrations/v32x/v320_m7_update_data.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ *
+ * PayPal Donation extension for the phpBB Forum Software package.
+ *
+ * @copyright (c) 2015 Skouat
+ * @license GNU General Public License, version 2 (GPL-2.0)
+ *
+ */
+
+namespace skouat\ppde\migrations\v32x;
+
+class v320_m7_update_data extends \phpbb\db\migration\migration
+{
+	public static function depends_on()
+	{
+		return ['\skouat\ppde\migrations\v32x\v320_m6_update_data'];
+	}
+
+	public function update_data()
+	{
+		return [
+			// INTL Settings
+			['config.add', ['ppde_default_locale', '']],
+			['config.add', ['ppde_intl_detected', false]],
+			['config.add', ['ppde_intl_version', '']],
+			['config.add', ['ppde_intl_version_valid', false]],
+			['config.update', ['ppde_first_start', true]],
+		];
+	}
+}